requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
